Splitting inputs to new file
