+++ src/common.h
@@ -53,8 +53,6 @@ char * objectFromPath(const char* myStr);
 void wPrintLine(int line, int col, wchar_t *textString);
 void printLine(int line, int col, char *textString);
 void topLineMessage(const char *message);
-int wReadLine(wchar_t *buffer, int buflen, wchar_t *oldbuf);
-int readline(char *buffer, int buflen, char *oldbuf);
 int check_dir(char *pwd);
 int check_file(char *file);
 void clear_workspace();
