+++ src/input.h
@@ -1,0 +1,2 @@
+int wReadLine(wchar_t *buffer, int buflen, wchar_t *oldbuf);
+int readline(char *buffer, int buflen, char *oldbuf);
