+++ src/sfmenus.c
@@ -27,6 +27,7 @@
 #include "common.h"
 #include "sf.h"
 #include "colors.h"
+#include "input.h"
 
 int c;
 int * pc = &c;
